requirements.txt
# Core AI & Deep Learning
# torch==2.8.0
# torchvision==0.23.0
transformers==5.4.0
accelerate==1.13.0
qwen-vl-utils==0.0.14

# Vector Database & Math
faiss-cpu==1.12.0
numpy==2.4.3

# Video & Image Processing
opencv-python==4.12.0.88
scenedetect==0.6.7.1
Pillow==11.3.0

# Web UI & CLI Utilities
streamlit==1.56.0
tqdm==4.67.1